Added api endpoints

+++ src/app/api/blog/[id]/route.ts
@@ -1,26 +1,42 @@
 import uploadImageOnCloudinary from "@/app/lib/imageUpload";
 import dbConnect from "@/lib/connectDb";
 import BlogModel from "@/models/blogModel";
+import { cookies } from "next/headers";
 import { NextRequest, NextResponse } from "next/server";
+import jwt from "jsonwebtoken";
+import UserModel from "@/models/userModel";
 
 export async function DELETE(req: NextRequest): Promise<NextResponse> {
     const id = req.nextUrl.pathname.split('/').pop();
     console.log(req.nextUrl.pathname);
     await dbConnect();
     try {
-        const deleteBlog = await BlogModel.findByIdAndDelete(id);
+        const cookieStore = await cookies();
+        const token = cookieStore.get("token")?.value; // Extract the token value
 
-        if (!deleteBlog) {
+        if (!token) {
+            return NextResponse.json({ message: "No token found, please login" }, { status: 401 });
+        }
+        
+        const decoded = jwt.verify(token, process.env.JWT_SECRET!) as { userId: string };
+        
+        const user = await UserModel.findById(decoded.userId);
+        const blog = await BlogModel.findById(id);
+
+        if (user?._id!.toString() === blog?.author.toString()) {
+            await BlogModel.findByIdAndDelete(id);
             return NextResponse.json({
-                success: false,
-                message: "Error while deleting the post"
-            }, { status: 400 });
-        };
+                message: "Post has been deleted successfully",
+                success: true
+            }, { status: 200 });
+        }
+
 
         return NextResponse.json({
-            success: true,
-            message: "Post has been deleted successfully"
-        }, { status: 200 });
+            success: false,
+            message: "Error while deleting the post"
+        }, { status: 401 });
+
     } catch (error) {
         console.log(error);
         return NextResponse.json({
@@ -53,7 +69,7 @@ export async function PUT(req: NextRequest): Promise<NextResponse> {
         }
 
         // Upload the file to Cloudinary
-        const multerFile = file as unknown as Express.Multer.File; // This may need more validation
+        const multerFile = file as unknown as Express.Multer.File;
         const imageUrl = await uploadImageOnCloudinary(multerFile);
 
         post.title = title || post.title;
@@ -73,4 +89,33 @@ export async function PUT(req: NextRequest): Promise<NextResponse> {
             success: false
         }, { status: 500 });
     }
+}
+
+
+export async function GET(req: NextRequest): Promise<NextResponse> {
+    await dbConnect();
+    console.log(req.nextUrl.pathname);
+    const id = req.nextUrl.pathname.split('/').pop();
+
+    try {
+        const blog = await BlogModel.findById(id);
+        if (!blog) {
+            return NextResponse.json({
+                message: "Blog not found",
+                sucess: false
+            }, { status: 401 });
+        }
+
+        return NextResponse.json({
+            blog,
+            success: true,
+        }, { status: 200 });
+    } catch (error) {
+        console.log(error);
+        return NextResponse.json({
+            message: "Internal server error",
+            success: false
+        }, { status: 500 });
+    }
+
 }
